Changed API and finished script for HSN Club
Changed the API in use to gspread and completed a version of the progrma that at the moment only work for the WWP HSN Chess Club. A more interactive script for other clubs uses will be implemented soon, and a column that displays how people moved up and down every month will be added.